recursiveloader: Do not recur into directories when file was expected
When verifying, do not let os.walk() recur into a directory for which we
have an entry requesting it to be a file. The verification is going
to fail on the directory entry anyway, so there is no point in seeking
matches further and this could have currently resulted in crossing
filesystem boundaries (since the cross-fs test is deferred to verifying
thread).