recursiveloader: Check for cross-device directories earlier in loop
Check for cross-device directory when starting to process the directory
rather than for each subdirectory to be processed. This will avoid
double stat() call in the future. It also means that cross-device
exceptions are raised a little later, when recurring into the directory
rather than before doing that. However, the starting directory is now
also tested.