hash: Read small-ish files into memory via a single call
Read files that are known to be small-ish (up to 1 MiB) into the memory
via a single read() call instead of reading in smaller parts. This is
mostly meant to workaround a bug in PyPy that causes wrong data when
small buffers are repeatedly passed to C extensions. Apparently slurping
the whole file does not exhibit this problem in gemato.